enhancement:  separation of components and organization for screens
